Remove circular dependency in schema and rdfa-utils
