+++ packages/ember-rdfa-editor/src/components/_private/rdfa-editor/property-editor/index.gts
@@ -3,7 +3,10 @@ import { tracked } from '@glimmer/tracking';
 import { addProperty } from '#root/commands/rdfa-commands/add-property.ts';
 import { removeProperty } from '#root/commands/rdfa-commands/remove-property.ts';
 import type { ResolvedPNode } from '#root/utils/_private/types.ts';
-import { isLinkTriple, type OutgoingTriple } from '#root/core/rdfa-processor.ts';
+import {
+  isLinkTriple,
+  type OutgoingTriple,
+} from '#root/core/rdfa-processor.ts';
 import {
   getSubjectsFromBacklinksOfRelationship,
   isLinkToNode,
