+++ packages/ember-rdfa-editor/src/plugins/rdfa-info/utils.ts
@@ -20,7 +20,7 @@ import type {
 } from '#root/utils/transaction-utils.ts';
 import TransformUtils from '#root/utils/_private/transform-utils.ts';
 import type { RemovePropertyArgs } from '#root/commands/rdfa-commands/remove-property.ts';
-import { isRdfaAttrs, type RdfaAttrs } from '#root/core/schema.ts';
+import { isRdfaAttrs, type RdfaAttrs } from '#root/core/rdfa-types.ts';
 import { unwrap } from '#root/utils/_private/option.ts';
 import MapUtils from '#root/utils/_private/map-utils.ts';
